Port javastraat Stats dBm center, farthest Null Island filter, hash flash.
Wire Avg Signal Quality centerLabel, reject (0,0)/implausible farthest
records, and resolve URL hash before scripts to avoid Dashboard flash.
Split Stats chart helpers to stay under the file-length cap.
