+++ docs/plans/v0.7.9-release.md
@@ -33,11 +33,11 @@ Much of J1/J2 is his work. Preserve and call it out; do not silently absorb.
 
 ### Wave J1 — javastraat core + small UI plucks
 
+- [x] Stats donut dBm + farthest 0,0
+- [x] Dashboard hash flash
 - [ ] MeshCore reconnect reliability
 - [ ] Serial live Region/preset/intervals/Bluetooth
 - [ ] MeshCore live radio + expanded presets
-- [ ] Stats donut dBm + farthest 0,0
-- [ ] Dashboard hash flash
 - [x] Battery `101` → "Powered" across cards/drawer/feed/chart (`29368c0`)
 - [x] Battery chip green/amber/red severity tiers (`5118d16`)
 - [x] Configuration GET failures toast (`676f7e3`)
